A new NewTab extension for Chrome.
The index.html file contains all the code for the actual project,
located at newtab.hoegendevelopments.com. The test file under packages
has an iframe that connects to the online version of the new tab,
allowing you to use it as an extension. To package it, just pack the
manifest and test.html together.